# swagger-ui/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/widgetmeta": {
            "put": {
                "description": "Update by json WidgetMeta",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WidgetMeta"
                ],
                "summary": "Update WidgetMeta",
                "parameters": [
                    {
                        "description": "Update WidgetMeta",
                        "name": "WidgetMeta",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.WidgetMeta"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "create WidgetMeta",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WidgetMeta"
                ],
                "summary": "Create WidgetMeta",
                "parameters": [
                    {
                        "description": "create WidgetMeta",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.WidgetMeta"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/widgetmeta/{id}": {
            "get": {
                "description": "Get WidgetMeta by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WidgetMeta"
                ],
                "summary": "Get all WidgetMeta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "WidgetMeta ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.WidgetMeta"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete WidgetMeta",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WidgetMeta"
                ],
                "summary": "Get all WidgetMeta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "WidgetMeta ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/widgetmetas/search": {
            "get": {
                "description": "get top 100 records WidgetMeta",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WidgetMeta"
                ],
                "summary": "Get all WidgetMeta",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "starting row",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "no of row limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.WidgetMeta"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.WidgetMeta": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "createdAt",
                    "type": "string"
                },
                "data": {
                    "description": "inputJson",
                    "type": "string"
                },
                "division": {
                    "description": "pageName",
                    "type": "string"
                },
                "id": {
                    "description": "WidgetName",
                    "type": "string"
                },
                "isActive": {
                    "description": "isActive",
                    "type": "boolean"
                },
                "module": {
                    "description": "datasource",
                    "type": "string"
                },
                "name": {
                    "description": "WidgetName",
                    "type": "string"
                },
                "page": {
                    "description": "datasourceType",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "updatedAt",
                    "type": "string"
                }
            }
        }
    }
}